@@ -26,7 +26,7 @@ group = 'com.github.tomakehurst'
 project.ext {
   versions = [
     handlebars     : '4.3.0',
-    jetty          : '9.4.48.v20220622',
+    jetty          : '11.0.11',
     guava          : '31.1-jre',
     jackson        : '2.13.4',
     xmlUnit        : '2.9.0',
@@ -45,9 +45,9 @@ dependencies {
   api "org.eclipse.jetty.http2:http2-server"
   api "org.eclipse.jetty:jetty-alpn-server"
   api "org.eclipse.jetty:jetty-alpn-java-server"
-  api "org.eclipse.jetty:jetty-alpn-openjdk8-server"
+  api "org.eclipse.jetty:jetty-alpn-server"
   api "org.eclipse.jetty:jetty-alpn-java-client"
-  api "org.eclipse.jetty:jetty-alpn-openjdk8-client"
+  api "org.eclipse.jetty:jetty-alpn-client"
 
   api "com.google.guava:guava:$versions.guava"
   api platform("com.fasterxml.jackson:jackson-bom:$versions.jackson")
@@ -82,9 +82,6 @@ dependencies {
     exclude group: 'org.mozilla', module: 'rhino'
   }
 
-  api 'commons-fileupload:commons-fileupload:1.4',  {
-    exclude group: 'commons-io', module: 'commons-io'
-  }
   api "commons-io:commons-io:2.11.0"
 
   testImplementation "junit:junit:4.13"
@@ -110,8 +107,13 @@ dependencies {
   testImplementation "org.eclipse.jetty:jetty-client"
   testImplementation "org.eclipse.jetty.http2:http2-http-client-transport"
 
+<<<<<<< HEAD
   testImplementation "ch.qos.logback:logback-classic:1.3.0"
 
+=======
+  testRuntimeOnly "org.slf4j:log4j-over-slf4j:2.0.0"
+  testRuntimeOnly "ch.qos.logback:logback-classic:1.4.0"
+>>>>>>> ffeee5ae (Jetty 11 support)
   testRuntimeOnly files('src/test/resources/classpath file source/classpathfiles.zip', 'src/test/resources/classpath-filesource.jar')
 
   testImplementation ('net.jockx:littleproxy:1.1.3') {
@@ -196,19 +198,21 @@ allprojects {
 
   version = '2.34.0'
 
-  sourceCompatibility = 1.8
-  targetCompatibility = 1.8
+  sourceCompatibility = 11
+  targetCompatibility = 11
 
   compileJava {
     options.encoding = 'UTF-8'
 
     // silences warnings about compiling against `sun` packages
     options.compilerArgs += '-XDenableSunApiLintControl'
+    options.compilerArgs += '--add-exports=java.base/sun.security.x509=ALL-UNNAMED'
   }
 
   compileTestJava {
     options.encoding = 'UTF-8'
     options.compilerArgs += '-XDenableSunApiLintControl'
+    options.compilerArgs += '--add-exports=java.base/sun.security.x509=ALL-UNNAMED'
   }
 
   test {
@@ -284,7 +288,7 @@ shadowJar {
   relocate "net.sf", "wiremock.net.sf"
   relocate "com.github.jknack", "wiremock.com.github.jknack"
   relocate "org.antlr", "wiremock.org.antlr"
-  relocate "javax.servlet", "wiremock.javax.servlet"
+  relocate "jakarta.servlet", "wiremock.jakarta.servlet"
   relocate "org.checkerframework", "wiremock.org.checkerframework"
   relocate "org.hamcrest", "wiremock.org.hamcrest"
 
@@ -294,7 +298,7 @@ shadowJar {
 
   mergeServiceFiles()
 
-  exclude 'META-INF/maven/javax.servlet/**'
+  exclude 'META-INF/maven/jakarta.servlet/**'
   exclude 'module-info.class'
   exclude 'handlebars-*.js'
 }
