@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2021 Thomas Akehurst
+ * Copyright (C) 2011-2022 Thomas Akehurst
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,8 +15,8 @@
  */
 package com.github.tomakehurst.wiremock.common;
 
+import jakarta.servlet.ServletContext;
 import java.io.File;
-import javax.servlet.ServletContext;
 
 public class ServletContextFileSource extends AbstractFileSource {
 
