@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2021 Thomas Akehurst
+ * Copyright (C) 2014-2022 Thomas Akehurst
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.tomakehurst.wiremock.jetty9;
+package com.github.tomakehurst.wiremock.jetty;
 
 import static com.github.tomakehurst.wiremock.common.Exceptions.throwUnchecked;
-import static com.github.tomakehurst.wiremock.jetty9.JettyUtils.unwrapResponse;
+import static com.github.tomakehurst.wiremock.jetty.JettyUtils.unwrapResponse;
 
 import com.github.tomakehurst.wiremock.core.FaultInjector;
 import com.google.common.base.Charsets;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.net.Socket;
 import java.nio.channels.SocketChannel;
-import javax.servlet.http.HttpServletResponse;
-import org.eclipse.jetty.io.ChannelEndPoint;
+import org.eclipse.jetty.io.SelectableChannelEndPoint;
 import org.eclipse.jetty.server.HttpChannel;
 import org.eclipse.jetty.server.Response;
 import org.eclipse.jetty.util.BufferUtil;
@@ -85,7 +85,7 @@ public class JettyFaultInjector implements FaultInjector {
 
   private Socket socket() {
     HttpChannel httpChannel = response.getHttpOutput().getHttpChannel();
-    ChannelEndPoint ep = (ChannelEndPoint) httpChannel.getEndPoint();
+    SelectableChannelEndPoint ep = (SelectableChannelEndPoint) httpChannel.getEndPoint();
     return ((SocketChannel) ep.getChannel()).socket();
   }
 }
