@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2021 Thomas Akehurst
+ * Copyright (C) 2011-2022 Thomas Akehurst
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -197,7 +197,7 @@ public class StandaloneAcceptanceTest {
     WireMockResponse response = testClient.get("/json/34567/");
     assertThat(response.statusCode(), is(200));
     assertThat(response.content(), is("<blob>BLAB</blob>"));
-    assertThat(response.firstHeader("Content-Type"), is("application/xml"));
+    assertThat(response.firstHeader("Content-Type"), is("application/xml;charset=utf-8"));
   }
 
   @Test
